Eliminate warnings and pass options as a dictionary (#299)
* Work to reduce warnings

* Cleared more warnings

* Fixed issue with test case

* Add black config file

* Add the ability to pass options as a dict in calls to energy/gradients

* Fix bug and remove compiler flag

* Enable findiff test

* Fix bad test case

* Reconcile second findiff test case

* Update reference output.

* Address comments and add example of options passing

* Update manual, add example, and test

* Fix printing

+++ forte/sci/aci_build_F.cc
@@ -883,7 +883,7 @@ AdaptiveCI::get_excited_determinants_batch(SharedMatrix evecs, SharedVector eval
     int nruns = static_cast<int>(std::ceil(guess_mem / max_mem));
 
     double total_excluded = 0.0;
-    int nbin = nruns;
+    size_t nbin = nruns;
     outfile->Printf("\n  Setting nbin to %d based on estimated memory (%6.3f MB)", nbin, guess_mem);
 
     if (options_->get_int("ACI_NBATCH") > 0) {
@@ -994,7 +994,7 @@ det_hash<double> AdaptiveCI::get_bin_F_space(int bin, int nbin, double E0, Share
                                ? thread_id * bin_size
                                : (n_dets % n_threads) * (bin_size + 1) +
                                      (thread_id - (n_dets % n_threads)) * bin_size;
-        int end_idx = start_idx + bin_size;
+        size_t end_idx = start_idx + bin_size;
 
         // Loop over P space determinants
         // size_t guess_a = nalpha_ * (no_ - nalpha_);
@@ -1276,7 +1276,7 @@ AdaptiveCI::get_bin_F_space_vecsort(int bin, int nbin, SharedMatrix evecs,
                                ? thread_id * bin_size
                                : (n_dets % n_threads) * (bin_size + 1) +
                                      (thread_id - (n_dets % n_threads)) * bin_size;
-        int end_idx = start_idx + bin_size;
+        size_t end_idx = start_idx + bin_size;
 
         // Loop over P space determinants
         // size_t guess_a = nalpha_ * (no_ - nalpha_);
